DF-14093 fix warnign as error
Former-commit-id: 89c1723cb78f8a7efdce6a6a2498f1497f130f69

+++ Sources/Internal/DLCManager/Private/DLCManagerImpl.cpp
@@ -1308,7 +1308,7 @@ DLCManager::Info DLCManagerImpl::GetInfo() const
     {
         info.infoCrc32 = initFooterOnServer.infoCrc32;
         info.metaCrc32 = initFooterOnServer.metaDataCrc32;
-        info.totalFiles = meta->GetFileCount();
+        info.totalFiles = static_cast<uint32>(meta->GetFileCount());
     }
     return info;
 }
